+++ api/database/migrations/2021_04_02_105313_create_order_deliveries_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
             $table->smallInteger('floor', unsigned: true)->nullable();
             $table->smallInteger('apartment', unsigned: true)->nullable();
             $table->boolean('service_to_door')->default(false);
-            $table->decimal('price', unsigned: true)->comment('Цена доставки');
+            $table->float('price', unsigned: true)->comment('Цена доставки');
 
             $table->foreignId('order_id')->constrained()->onUpdate('cascade')->onDelete('cascade');
             $table->foreignId('location_id')->nullable()->constrained()->onUpdate('cascade')->onDelete('set null');
