+++ api/database/migrations/2022_03_04_071406_create_product_statistics_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->integer('orders', unsigned: true)->default(0);
             $table->integer('reviews', unsigned: true)->default(0);
             $table->integer('cancellations', unsigned: true)->default(0);
-            $table->decimal('rating', unsigned: true)->default(0);
+            $table->float('rating', unsigned: true)->default(0);
             $table->timestamps();
 
             $table->primary('id');
